Reference API

Documentation for StereoSSAM.jl's public interface.

Index

StereoSSAM.IO.readstereoseq

StereoSSAM.IO.readstereoseqbinned

StereoSSAM.KDE.assigncelltype

StereoSSAM.KDE.gaussiankernel

StereoSSAM.KDE.kde

StereoSSAM.LocalMax.findlocalmaxima

StereoSSAM.LocalMax.getlocalmaxima

StereoSSAM.Utils.crop!

StereoSSAM.Utils.mask!

StereoSSAM.Utils.totalrna

API

StereoSSAM.IO.readstereoseq — Method

readstereoseq(file)

Read StereoSeq file as vector of SparseMatrixCSC.

StereoSSAM.IO.readstereoseqbinned — Method

readstereoseqbinned(file, s::Integer)

Read StereoSeq file and aggregate locations with bin size s.

StereoSSAM.Utils.crop! — Method

crop!(counts, slice)

Crop the counts to slice.

StereoSSAM.Utils.mask! — Method

mask!(counts, mask::AbstractMatrix{Bool})

Remove all counts for which mask is false.

StereoSSAM.Utils.totalrna — Method

totalrna(counts)

Caclulate the totalrna as sum of all genes for each pixel.

StereoSSAM.KDE.assigncelltype — Method

function assigncelltype(
    counts::DimArray{T,1}, signatures::AbstractDataFrame, kernel; celltypes=nothing
) where {T<:AbstractMatrix}

Assign a celltype to each pixel.

The cosine similarity is calculated using signatures to assign the celltype with the highest similarity to each pixel.

The eltype(kernel) will be used for calculations and signatures will be cast to it.

Arguments

signatures::AbstractDataFrame: celltype signatures (celltypes x genes).

celltypes::Vector{AbstractString}=nothing: celltype names.

StereoSSAM.KDE.gaussiankernel — Method

gaussiankernel(σ::Real, r::Real)

Generate a gaussian kernel with bandwidth σ and radius r*σ

StereoSSAM.KDE.kde — Method

kde(counts::AbstractArray{T}, kernel) where {T<:Real}

Calculate kernel density estimate.

StereoSSAM.LocalMax.findlocalmaxima — Method

findlocalmaxima(img, mindist::Integer; threshold::Real=0)

Find local maxima of the img.

The input should be a kde of the totalrna or comparable.

StereoSSAM.LocalMax.getlocalmaxima — Method

getlocalmaxima(counts, localmax, kernel; genes=nothing)

Load KDE with kernel for coordinates at localmax.

Arguments

genes=nothing: vector of genes for which to calculate KDE.