Reference API

Documentation for Sainsc.jl's public interface.

Index

Sainsc.GridCount.AbstractGridCounts

Sainsc.GridCount.GridCounts

Sainsc.GridCount.GridCounts

Sainsc.GridCount.GridCounts

Sainsc.GridCount.GridCounts

Sainsc.GridCount.crop!

Sainsc.GridCount.gridsize

Sainsc.GridCount.mask!

Sainsc.GridCount.totalrna

Sainsc.IO.readGEMfile

Sainsc.IO.readstereoseq

Sainsc.IO.readstereoseqbinned

Sainsc.KDE.assigncelltype

Sainsc.KDE.gaussiankernel

Sainsc.KDE.kde

Sainsc.LocalMax.findlocalmaxima

Sainsc.LocalMax.getlocalmaxima

API

Sainsc.GridCount.AbstractGridCounts — Type

AbstractGridCounts

A stack of genes each consisting of a 2D count array of the same size.

Sainsc.GridCount.GridCounts — Type

GridCounts{G,T<:Number} <: AbstractGridCounts

A stack of genes::G each consisting of a 2D count::T array of the same size.

Sainsc.GridCount.GridCounts — Method

GridCounts(df::DataFrame, binsize::Integer)

Construct from a DataFrame with columns 'x' (Real), 'y' (Real), 'count', and 'geneID' (Pool) binning the data by binsize.

Sainsc.GridCount.GridCounts — Method

GridCounts(df::DataFrame)

Construct from a DataFrame with columns 'x' (Integer), 'y' (Integer), 'count', and 'geneID' (Pool).

Sainsc.GridCount.GridCounts — Method

GridCounts(counts::Dict)

All values of the dict must be arrays of the same size.

Sainsc.GridCount.crop! — Method

crop!(counts, slice)

Crop each gene layer in counts by indexing with slice.

Sainsc.GridCount.gridsize — Method

gridsize(counts)

The size of the grid of each layer.

Sainsc.GridCount.mask! — Method

mask!(counts, mask::AbstractMatrix{Bool})

Remove all counts in each gene layer for which mask is false.

Sainsc.GridCount.totalrna — Method

totalrna(counts)

Caclulate the totalrna as sum of all genes for each pixel.

Sainsc.IO.readGEMfile — Method

readGEMfile(file)

Read file in GEM format as DataFrames.DataFrame.

Sainsc.IO.readstereoseq — Method

readstereoseq(file)

Read StereoSeq file as GridCounts.

Sainsc.IO.readstereoseqbinned — Method

readstereoseqbinned(file, binsize::Integer)

Read StereoSeq file and aggregate into bins.

Sainsc.KDE.assigncelltype — Method

assigncelltype(counts, signatures, kernel; celltypes=nothing, log=false)

Assign a celltype to each pixel.

The cosine similarity is calculated using signatures to assign the celltype with the highest similarity to each pixel.

The eltype(kernel) will be used for calculations and signatures will be cast to it.

Arguments

signatures::AbstractDataFrame: celltype signatures (celltypes x genes).

celltypes::Vector{AbstractString}=nothing: optional celltype names.

log::Bool: whether to log-transform the KDE. Useful if signatures are calculated from log-transformed gene expression.

Sainsc.KDE.gaussiankernel — Method

gaussiankernel(σ::Real, r::Real)

Generate a gaussian kernel with bandwidth σ and radius r * σ

Sainsc.KDE.kde — Method

kde(counts, kernel)

Calculate kernel density estimate.

Arguments

kernel: usually a centered OffsetArrays.OffsetArray.

Sainsc.LocalMax.findlocalmaxima — Method

findlocalmaxima(img, mindist::Integer; threshold::Real=0)

Find local maxima of the img.

The input should be a kde of the totalrna or comparable.

Sainsc.LocalMax.getlocalmaxima — Method

getlocalmaxima(counts, localmax, kernel; genes=nothing)

Load KDE with kernel for coordinates at localmax.

Arguments

localmax: vector of locations (indices).

genes=nothing: vector of genes for which to calculate KDE. If nothing all genes are used.